Refine StoreDB tests
Create isolated test file. Prevent unit dependency acrossed tests, specifically save methods
