--- tests/directories.json
@@ -1,47 +1,0 @@
-[
-    {
-        "root": "foo",
-        "last_seen": 100,
-        "file_count": 10
-    },
-    {
-        "root": "foo",
-        "last_seen": 110,
-        "file_count": 20
-    },
-    {
-        "root": "foo",
-        "last_seen": 120,
-        "file_count": 30
-    },
-    {
-        "root": "bar",
-        "last_seen": 100,
-        "file_count": 10
-    },
-    {
-        "root": "bar",
-        "last_seen": 120,
-        "file_count": 20
-    },
-    {
-        "root": "bar",
-        "last_seen": 130,
-        "file_count": 30
-    },
-    {
-        "root": "baz",
-        "last_seen": 100,
-        "file_count": 10
-    },
-    {
-        "root": "baz",
-        "last_seen": 120,
-        "file_count": 20
-    },
-    {
-        "root": "baz",
-        "last_seen": 130,
-        "file_count": 30
-    }
-]
