--- tests/files.json
@@ -1,26 +1,0 @@
-[
-    {
-        "root": "some/path/here",
-        "filename": "foo",
-        "first_seen": 10,
-        "last_seen": 10,
-        "age_seconds": 0,
-        "removed": 0
-    },
-    {
-        "root": "some/path/here",
-        "filename": "bar",
-        "first_seen": 10,
-        "last_seen": 10,
-        "age_seconds": 0,
-        "removed": 0
-    },
-    {
-        "root": "some/path/here",
-        "filename": "baz",
-        "first_seen": 10,
-        "last_seen": 10,
-        "age_seconds": 0,
-        "removed": 0
-    }
-]
