+++ tests/walkwatcher_test.py
@@ -1,46 +1,9 @@
 from __future__ import annotations
 
-import json
-from datetime import datetime
-from pathlib import Path
-from random import shuffle
-from unittest.mock import patch
-
 import pytest
 
 from walk_watcher.walkwatcher import Directory
 from walk_watcher.walkwatcher import File
-from walk_watcher.walkwatcher import StoreDB
-
-DIRECTORIES_FILE = Path(__file__).parent / "directories.json"
-FILES_FILE = Path(__file__).parent / "files.json"
-MAX_IS_RUNNING_AGE = 5 * 60  # 5 minutes
-
-
-@pytest.fixture
-def store_db() -> StoreDB:
-    return StoreDB(
-        ":memory:",
-        max_is_running_age=MAX_IS_RUNNING_AGE,
-    )
-
-
-@pytest.fixture
-def directories() -> list[Directory]:
-    """Provide a shuffled list of directories."""
-    with DIRECTORIES_FILE.open() as f:
-        directories = [Directory(**d) for d in json.load(f)]
-    shuffle(directories)
-    return directories
-
-
-@pytest.fixture
-def files() -> list[File]:
-    """Provide a shuffled list of files."""
-    with FILES_FILE.open() as f:
-        files = [File(**d) for d in json.load(f)]
-    shuffle(files)
-    return files
 
 
 def test_model_directory_str() -> None:
@@ -106,296 +69,3 @@ def test_model_file_as_metric_line_raises_on_invalid_metric_name() -> None:
     file = File("/foo/bar", "baz.txt", 1234567890, 1234568190, 0, 0)
     with pytest.raises(ValueError):
         file.as_metric_line("walk_watcher test")
-
-
-def test_storedb_works_with_context_manager() -> None:
-    with patch("walk_watcher.walkwatcher.StoreDB.start_run") as start_run_mock:
-        with patch("walk_watcher.walkwatcher.StoreDB.stop_run") as stop_run_mock:
-            with StoreDB(":memory:") as store_db:
-                store_db.start_run()
-                store_db.stop_run()
-
-    assert start_run_mock.call_count == 2
-    assert stop_run_mock.call_count == 2
-
-
-def test_create_file_table(store_db: StoreDB) -> None:
-    cursor = store_db._connection.cursor()
-    cursor.execute("SELECT name FROM sqlite_master WHERE type='table'")
-    tables = [name[0] for name in cursor.fetchall()]
-
-    assert "files" in tables
-
-
-def test_create_directory_table(store_db: StoreDB) -> None:
-    cursor = store_db._connection.cursor()
-    cursor.execute("SELECT name FROM sqlite_master WHERE type='table'")
-    tables = [name[0] for name in cursor.fetchall()]
-
-    assert "directories" in tables
-
-
-def test_create_system_table(store_db: StoreDB) -> None:
-    cursor = store_db._connection.cursor()
-    cursor.execute("SELECT name FROM sqlite_master WHERE type='table'")
-    tables = [name[0] for name in cursor.fetchall()]
-
-    assert "system" in tables
-
-
-def test_system_table_is_saved_on_init(store_db: StoreDB) -> None:
-    cursor = store_db._connection.cursor()
-    cursor.execute("SELECT * FROM system")
-    rows = cursor.fetchall()
-
-    assert len(rows) == 1
-    assert rows[0][0] == ":memory:"  # database path is correct
-    assert rows[0][2] == 0  # is_running is False
-    # We don't check the last_seen timestamp for simplicity.
-
-
-def test_start_run_updates_system_table(store_db: StoreDB) -> None:
-    store_db.start_run()
-    cursor = store_db._connection.cursor()
-    cursor.execute("SELECT is_running FROM system")
-    rows = cursor.fetchone()
-
-    assert rows[0] == 1  # is_running is True
-
-
-def test_start_run_with_stale_is_running_flag_updates_system_table(
-    store_db: StoreDB,
-) -> None:
-    store_db.start_run()
-    cursor = store_db._connection.cursor()
-    new_timestamp = datetime.now().timestamp() - MAX_IS_RUNNING_AGE - 1
-    cursor.execute(
-        "UPDATE system SET last_run = ?",
-        (new_timestamp,),
-    )
-
-    store_db.start_run()
-
-    cursor.execute("SELECT last_run FROM system")
-    rows = cursor.fetchone()
-
-    assert rows[0] != new_timestamp  # last_seen was updated
-
-
-def test_start_run_while_running_raises(store_db: StoreDB) -> None:
-    store_db.start_run()
-
-    with pytest.raises(RuntimeError):
-        store_db.start_run()
-
-
-def test_end_run_updates_system_table(store_db: StoreDB) -> None:
-    cursor = store_db._connection.cursor()
-    cursor.execute("UPDATE system SET is_running = 1")
-
-    store_db.stop_run()
-    cursor.execute("SELECT is_running FROM system")
-    rows = cursor.fetchone()
-
-    assert rows[0] == 0  # is_running is False
-
-
-def test_save_directories(
-    store_db: StoreDB,
-    directories: list[Directory],
-) -> None:
-    # We save the directories twice to ensure that the
-    # UNIQUE(root, last_seen) constraint is working.
-    store_db.save_directories(directories)
-    store_db.save_directories(directories)
-    cursor = store_db._connection.cursor()
-    cursor.execute("SELECT * FROM directories")
-    rows = cursor.fetchall()
-
-    assert len(rows) == len(directories)
-    assert all(
-        row[1:] == (directory.root, directory.last_seen, directory.file_count)
-        for row, directory in zip(rows, directories)
-    )
-
-
-def test_save_files_empty_rows(
-    store_db: StoreDB,
-    files: list[File],
-) -> None:
-    store_db.save_files(files)
-    cursor = store_db._connection.cursor()
-    cursor.execute("SELECT * FROM files")
-    rows = cursor.fetchall()
-
-    assert len(rows) == len(files)
-    assert all(
-        row[1:]
-        == (
-            file.root,
-            file.filename,
-            file.last_seen,
-            file.last_seen,
-            file.age_seconds,
-            0,
-        )
-        for row, file in zip(rows, files)
-    )
-
-
-def test_save_file_existing_row_updated(
-    store_db: StoreDB,
-    files: list[File],
-) -> None:
-    store_db.save_files(files)
-
-    # Replace last with new: the same root and filename but 300 seconds older.
-    expected_first_seen = files[-1].first_seen
-    new_age = 600
-    files[-1] = File(
-        root=files[-1].root,
-        filename=files[-1].filename,
-        last_seen=files[-1].first_seen + new_age,
-    )
-
-    store_db.save_files(files)
-
-    cursor = store_db._connection.cursor()
-    cursor.execute("SELECT * FROM files")
-    last_row = cursor.fetchall()[-1]
-    print(last_row)
-
-    assert last_row[1:] == (
-        files[-1].root,
-        files[-1].filename,
-        expected_first_seen,
-        files[-1].last_seen,
-        new_age,
-        0,
-    )
-
-
-def test_save_file_add_new_row_with_existing_rows(
-    store_db: StoreDB,
-    files: list[File],
-) -> None:
-    store_db.save_files(files)
-
-    # Add a new file with a different root and filename
-    new_file = File(
-        root="new_root",
-        filename="new_file",
-        first_seen=1234567890,
-        last_seen=1234567890,
-        age_seconds=0,
-        removed=0,
-    )
-    files.append(new_file)
-
-    store_db.save_files(files)
-
-    cursor = store_db._connection.cursor()
-    cursor.execute("SELECT * FROM files")
-    last_row = cursor.fetchall()[-1]
-
-    assert last_row[1:] == (
-        new_file.root,
-        new_file.filename,
-        new_file.first_seen,
-        new_file.last_seen,
-        new_file.age_seconds,
-        new_file.removed,
-    )
-
-
-def test_save_files_marks_all_removed(
-    store_db: StoreDB,
-    files: list[File],
-) -> None:
-    store_db.save_files(files)
-
-    # Save an empty list of files
-    store_db.save_files([])
-
-    cursor = store_db._connection.cursor()
-    cursor.execute("SELECT removed FROM files")
-    removed = [value[0] for value in cursor.fetchall()]
-
-    assert all(removed)
-
-
-def test_get_directory_rows(store_db: StoreDB) -> None:
-    directories = [
-        Directory(root="root1", last_seen=1234567891, file_count=0),
-        Directory(root="root1", last_seen=1234567890, file_count=0),
-    ]
-
-    store_db.save_directories(directories)
-
-    rows = store_db.get_directory_rows()
-
-    assert len(rows) == 1
-    assert rows[0] == directories[0]
-
-
-def test_get_file_rows(store_db: StoreDB) -> None:
-    files = [
-        File(
-            root="root1",
-            filename="file1",
-            first_seen=1234567891,
-            last_seen=1234567891,
-            age_seconds=0,
-            removed=0,
-        ),
-        File(
-            root="root1",
-            filename="file1",
-            first_seen=1234567891,
-            last_seen=1234567891,
-            age_seconds=0,
-            removed=1,
-        ),
-        File(
-            root="root2",
-            filename="file2",
-            first_seen=1234567890,
-            last_seen=1234567891,
-            age_seconds=0,
-            removed=0,
-        ),
-    ]
-    directory = Directory(root="root1", last_seen=1234567891, file_count=0)
-
-    store_db.save_files(files)
-
-    rows = store_db.get_file_rows(directory)
-
-    assert len(rows) == 1
-    assert rows[0] == files[0]
-
-
-def test_get_oldest_files(store_db: StoreDB) -> None:
-    init_files = [
-        File("root1", "file1", 1234567890),
-        File("root1", "file2", 1234567890),
-        File("root2", "file1", 1234567890),
-        File("root2", "file2", 1234567890),
-    ]
-    store_db.save_files(init_files)
-    update_files = [
-        File("root1", "file1", 1234567893),
-        File("root1", "file2", 1234567891),
-        File("root2", "file1", 1234567891),
-        File("root2", "file2", 1234567893),
-    ]
-    store_db.save_files(update_files)
-
-    rows = store_db.get_oldest_files()
-    print(rows)
-
-    assert len(rows) == 2
-    assert rows[0].root == "root1"
-    assert rows[0].filename == "file1"
-    assert rows[1].root == "root2"
-    assert rows[1].filename == "file2"
